requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.2.0
async-timeout==3.0.1
attrs==21.4.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.11
click==8.0.4
colorama==0.4.4
discord==1.7.3
Flask==2.0.3
frozenlist==1.3.0
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
multidict==6.0.2
Pillow==9.0.1
praw==7.5.0
prawcore==2.3.0
requests==2.27.1
typing_extensions==4.0.1
update-checker==0.18.0
urllib3==1.26.8
wavelink==1.1.1
websocket-client==1.2.3
Werkzeug==2.0.3
yarl==1.7.2
git+https://github.com/Rapptz/discord.py